requirements.txt
Django==5.0.2
opencv-python-headless>=4.9.0.80
numpy>=1.24.3
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
mediapipe>=0.10.8
scikit-learn>=1.3.2
setuptools>=69.0.3 